backend/requirements.txt
# FastAPI and Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.2
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0
geoalchemy2==0.14.2

# Time Series Database
timescaledb==0.1.0

# Data Processing
pandas==2.1.4
numpy==1.26.2
polars==0.19.19

# ML/AI
scikit-learn==1.3.2
tensorflow==2.15.0
torch==2.1.1
xgboost==2.0.3
lightgbm==4.1.0
joblib==1.3.2

# Data Validation
great-expectations==0.18.8
pandera==0.18.0

# API Clients
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# Message Queue & Streaming
kafka-python==2.0.2
aiokafka==0.10.0
pika==1.3.2
redis==5.0.1
celery==5.3.4

# IoT Protocols
paho-mqtt==1.6.1
asyncua==1.0.6

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==41.0.7
cryptography[fernet]

# Monitoring & Logging
prometheus-client==0.19.0
structlog==23.2.0
python-json-logger==2.0.7
psutil==5.9.6

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
click==8.1.7
tqdm==4.66.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.25.2
faker==20.1.0

# Code Quality
black==23.12.1
flake8==6.1.0
mypy==1.7.1
isort==5.12.0

# Data Sources
kaggle==1.5.16
boto3==1.34.0

# Feature Store
feast==0.36.1

# Model Serving
mlflow==2.8.1
